feat: expose classes of package `audioplayers_platform_interface`
