+++ packages/audioplayers/lib/audioplayers.dart
@@ -1,11 +1,11 @@
-export 'package:audioplayers_platform_interface/api/audio_context.dart';
-export 'package:audioplayers_platform_interface/api/audio_context_config.dart';
-export 'package:audioplayers_platform_interface/api/for_player.dart';
-export 'package:audioplayers_platform_interface/api/log_level.dart';
-export 'package:audioplayers_platform_interface/api/player_mode.dart';
-export 'package:audioplayers_platform_interface/api/player_state.dart';
-export 'package:audioplayers_platform_interface/api/release_mode.dart';
-export 'package:audioplayers_platform_interface/global_platform_interface.dart';
+export 'package:audioplayers_platform_interface/src/api/audio_context.dart';
+export 'package:audioplayers_platform_interface/src/api/audio_context_config.dart';
+export 'package:audioplayers_platform_interface/src/api/for_player.dart';
+export 'package:audioplayers_platform_interface/src/api/log_level.dart';
+export 'package:audioplayers_platform_interface/src/api/player_mode.dart';
+export 'package:audioplayers_platform_interface/src/api/player_state.dart';
+export 'package:audioplayers_platform_interface/src/api/release_mode.dart';
+export 'package:audioplayers_platform_interface/src/global_platform_interface.dart';
 
 export 'src/audio_cache.dart';
 export 'src/audio_pool.dart';
