+++ packages/audioplayers/test/audioplayers_test.dart
@@ -1,5 +1,5 @@
 import 'package:audioplayers/audioplayers.dart';
-import 'package:audioplayers_platform_interface/method_channel_interface.dart';
+import 'package:audioplayers_platform_interface/src/method_channel_interface.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
