+++ packages/audioplayers_platform_interface/lib/src/api/audio_context_config.dart
@@ -1,6 +1,6 @@
 import 'dart:io' show Platform;
 
-import 'package:audioplayers_platform_interface/api/audio_context.dart';
+import 'package:audioplayers_platform_interface/src/api/audio_context.dart';
 
 /// This class contains flags to control several secondary, platform-specific
 /// aspects of audio playback, like how this audio interact with other audios,
