+++ packages/audioplayers_platform_interface/lib/src/api/for_player.dart
