+++ packages/audioplayers_platform_interface/lib/src/global_platform_interface.dart
@@ -1,6 +1,6 @@
-import 'package:audioplayers_platform_interface/api/audio_context.dart';
-import 'package:audioplayers_platform_interface/api/log_level.dart';
-import 'package:audioplayers_platform_interface/method_channel_interface.dart';
+import 'package:audioplayers_platform_interface/src/api/audio_context.dart';
+import 'package:audioplayers_platform_interface/src/api/log_level.dart';
+import 'package:audioplayers_platform_interface/src/method_channel_interface.dart';
 import 'package:flutter/services.dart';
 
 abstract class GlobalPlatformInterface {
