+++ packages/audioplayers_platform_interface/lib/src/streams_interface.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:audioplayers_platform_interface/api/for_player.dart';
+import 'package:audioplayers_platform_interface/src/api/for_player.dart';
 import 'package:flutter/foundation.dart';
 
 mixin StreamsInterface {
