+++ packages/audioplayers_platform_interface/test/logger_test.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
-import 'package:audioplayers_platform_interface/api/log_level.dart';
-import 'package:audioplayers_platform_interface/global_platform_interface.dart';
+import 'package:audioplayers_platform_interface/audioplayers_platform_interface.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
